+++ R/check_model.R
@@ -28,9 +28,10 @@
 #' m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
 #' check_model(m)
 #'
-#' library(lme4)
-#' m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
-#' check_model(m, panel = FALSE)
+#' if (require("lme4")) {
+#'   m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
+#'   check_model(m, panel = FALSE)
+#' }
 #'
 #' if (require("rstanarm")) {
 #'   m <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)