+++ NEWS.md
@@ -6,9 +6,12 @@
 
 ## New functions
 
-* `check_itemscale()` to describe properties of (sub)scales, based on results from principal component analyses.
 * `performance_aic()` as a small wrapper that returns the AIC. It is a generic function that also works for some models that don't have a AIC method (like Tweedie models).
 * `performance_lrt()` as a small wrapper around `anova()` to perform a Likelihood-Ratio-Test for model comparison.
+
+## Bug fixes
+
+* Fix issues with CRAN checks.
 
 ## Changes to functions