+++ NEWS.md
@@ -1,3 +1,9 @@
+# performance 0.4.4
+
+## New functions
+
+* `check_itemscale()` to describe various measures of internal consistencies for scales which were built from several items from a PCA, using `parameters::principal_components()`.
+
 # performance 0.4.3
 
 ## General